move part of the derive macro into a new trait
