+++ Cargo.toml
@@ -61,7 +61,7 @@ inherits = "release"
 lto = true
 
 [features]
-default = []
+default = ["pcsaft", "dft"]
 dft = ["feos-dft", "petgraph"]
 estimator = []
 association = []
