+++ feos-dft/src/lib.rs
@@ -18,7 +18,9 @@ mod weight_functions;
 
 pub use convolver::{Convolver, ConvolverFFT};
 pub use functional::{HelmholtzEnergyFunctional, MoleculeShape, DFT};
-pub use functional_contribution::{FunctionalContribution, FunctionalContributionDual};
+pub use functional_contribution::{
+    FunctionalContribution, FunctionalContributionDual, PartialDerivativesDual,
+};
 pub use geometry::{Axis, Geometry, Grid};
 pub use profile::{DFTProfile, DFTSpecification, DFTSpecifications};
 pub use solver::{DFTSolver, DFTSolverLog};
