+++ src/pcsaft/dft/dispersion.rs
@@ -5,8 +5,8 @@ use crate::pcsaft::eos::dispersion::{A0, A1, A2, B0, B1, B2};
 use feos_core::EosResult;
 use feos_derive::FunctionalContribution;
 use feos_dft::{
-    FunctionalContribution, FunctionalContributionDual, WeightFunction, WeightFunctionInfo,
-    WeightFunctionShape,
+    FunctionalContribution, FunctionalContributionDual, PartialDerivativesDual, WeightFunction,
+    WeightFunctionInfo, WeightFunctionShape,
 };
 use ndarray::*;
 use num_dual::*;
