+++ src/pcsaft/dft/hard_chain.rs
@@ -3,8 +3,8 @@ use crate::hard_sphere::HardSphereProperties;
 use feos_core::EosResult;
 use feos_derive::FunctionalContribution;
 use feos_dft::{
-    FunctionalContribution, FunctionalContributionDual, WeightFunction, WeightFunctionInfo,
-    WeightFunctionShape,
+    FunctionalContribution, FunctionalContributionDual, PartialDerivativesDual, WeightFunction,
+    WeightFunctionInfo, WeightFunctionShape,
 };
 use ndarray::*;
 use num_dual::*;
